# ClaimMessages/MyClaim.cs
using System;

namespace ClaimMessages
{
    public class MyClaim
    {
        public string ClaimId { get; set; }
        public string ClientName { get; set; }
        public string WorkName { get; set; }
        public string ReferenceNum { get; set; }
    }
}