# CmsIntegrationApi/swagger.json
{
  "x-generator": "NSwag v12.0.15.0 (NJsonSchema v9.13.22.0 (Newtonsoft.Json v11.0.0.0))",
  "swagger": "2.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "host": "localhost:8202",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json-patch+json",
    "application/json",
    "text/json",
    "application/*+json"
  ],
  "paths": {
    "/api/Claims": {
      "get": {
        "tags": [
          "Claims"
        ],
        "operationId": "Claims_GetAll",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/MyClaim"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Claims"
        ],
        "operationId": "Claims_Post",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "name": "claim",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/MyClaim"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/api/Claims/{id}": {
      "get": {
        "tags": [
          "Claims"
        ],
        "operationId": "Claims_Get",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true,
            "x-nullable": true
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/MyClaim"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "MyClaim": {
      "type": "object",
      "additionalProperties": false,
      "properties": {
        "claimId": {
          "type": "string"
        },
        "clientName": {
          "type": "string"
        },
        "workName": {
          "type": "string"
        },
        "referenceNum": {
          "type": "string"
        }
      }
    }
  }
}